Split status markdown run fixtures
